Develop에서 main으로 머지 (#55)
* Document: Github Template

Github Template 추가하였습니다
- Issue
- Pull Request

Issues: #1

* Core : 백엔드 프로젝트 환경 세팅

백엔드 프로젝트 생성 및 초기 환경 세팅
- Dependencies : Lombok, Spring Data JPA, Validation, Spring Web, H2 Database, MySQL Driver
- Java 17
- Gradle

Issues : #4

* Feature/entity 초기 설정/9 (#12)

* Feature : Member domain 추가

회원 도메인 추가

Issues: #9

* Feature: Submit(풀이) 도메인 추가

풀이 entity 초기 설정

Issues: #9

* Feature: Problem(문제) 도메인 추가

Problem entity 초기 설정

Issues:#9

* Feature: BaseEntity 추가

생성일, 수정일 전용 BaseEntity class 추가

Issues: #9

* Feature/setting ide/14-15 (#16)

* Core: 파싱 관련 Gson 라이브러리 추가.

Issues: #14

* Feature: Problem 엔티티 연관관계 추가

Issues: #14

* Feature: Testcase 엔티티 추가

Issues: #14

* Feature: Testcase Repository 추가

JpaRepository 로 생성
Issues: #14

* Feature: TestcaseService 생성

* 공개 테케 가져오기
* 히든 테케 가져오기
* 코드 실행
* 코드 제출
Issues: #14

* Feature: IdeController 생성

코드 실행, 제출 API 작성
Issues: #14

* Feature: Ide 관련 DTO 작성

Issues: #14

* Feature: compile builder 작성

Issues: #14

* Revert "Feature: IdeController 생성"

This reverts commit 3782b7651b1e2cc566d774456057444fb98e233c.

* Feature: Ide Controller 작성

코드 실행, 제출 API 작성
Issues: #15

* Feature/solution/13 (#17)

* Feature: BaseEntity 상속 추가, @Getter 애노테이션 추가

Issue: #13

* Feature: SolutionController 생성 및 풀이 관리 API 추가

제공 기능:
- 전체 정답 솔루션 조회: `/api/solutions`
- 문제별 정답 솔루션 조회: `/api/solutions/{problemId}`
- 솔루션 개수 조회: `/api/solutions/count`
- 페이지별 솔루션 조회: `/api/solutions/{problemId}/group`

Issues: #13

* Feature: SolutionDTO 생성

SolutionDTO 클래스를 생성하여 솔루션 관련 데이터 전송 객체를 정의하였습니다. 이 DTO는 아래 값들을 포함하고 있습니다.
 문제 ID, 회원 ID, 정답 여부, 제출 코드, 사용 언어, 문제 상태, 생성일, 수정일, 회원 이름, 프로필 이미지 링크

Issues: #13

* Feature: SolutionRepository 생성 및 커스텀 쿼리 메소드 추가

주요 메소드:
- `findByIsCorrectTrue()`: 정답인 모든 솔루션과 해당 회원 정보 조회
- `findByIsCorrectTrueAndProblemId(Long problemId)`: 특정 문제에 대한 정답 솔루션 조회
- `countByIsCorrectTrue()`: 정답인 솔루션 총 개수 조회
- `findByIsCorrectTrueAndLanguageAndMemberIdAndProblemId(String language, Long memberId, Long problemId, Pageable pageable)`: 언어, 회원 ID, 문제 ID를 기준으로 한 정답 솔루션 페이지 조회
- `findByIsCorrectTrueAndLanguageAndProblemId(String language, Long problemId, Pageable pageable)`: 특정 언어와 문제 ID 기준 정답 솔루션 페이지 조회

Issues: #13

* Feature: SolutionService 클래스 생성 및 풀이 조회 및 DTO 변환 기능 구현

구현된 주요 기능:
- `allCorrectSolutions()`: 모든 풀이 조회
- `allCorrectSpecificSolutions(Long problemId)`: 특정 문제에 대한 정답 풀이 조회
- `solutionCounts()`: 정답 풀이 총 개수 조회
- `findSolutionPages(int page, int size, String language, String type, Long problemId)`: 특정 조건에 따른 풀이 페이지 조회

Issues: #13

* Feature: testcase 관련 응답 

* Feat: IdeController API 추가

기본 테스트 케이스 값, 타입, 결과 불러오는 API 작성
Issues: #18

* Refactor: 쿼리 메소드로 변경

Before: Service에 로직 작성
After: 쿼리 메소드로 변경
Issues: #19

* Refactor: 쿼리메소드 작성

기존 서비스 로직 쿼리 메소드로 대체
Issues: #19

* Feat: Ide 관련 DTO 추가, 변경

-추가 : 테스트케이스 불러오는 API에 대한 ParamDto 추가
-변경 : 기존 DTO에 성공/실패 여부, 제출/실행 여부 추가

* Refactor: testcase response 변경

* Refactor: 테스트케이스 Response 형태 변경

Issues: #21

* Feature: 테스트케이스 DTO 추가

Issues: #21

* Feature : 구글로그인구현 #23 (#26)

구글 로그인 구현

* Feature : 채팅 구현 #24 (#28)

* Feature : openAi 설정 추가

Issues : #24

* Feature : 채팅 관련 dto 생성

유저가 보내는 메세지 request,  gpt에서 오는 응답 response 를 생성. 프로그램에서 주는 response 구현 필요.

#24

* Feature : gpt에게 채팅 보내기 구현

gpt api를 사용하여 채팅 보내고 응답 받을 수 있음. prompt 설계나 db 저장은 미구현.

Issues : #24

* Feature : 채팅 요청 controller 생성

Issues : #24

* Feature : gpt prompt 설계를 위한 코드 추가

토큰 제한과 system 설정을 추가했다.

#24

* Feature : 배포용 application 설정 및 docker 파일 작성

* Core : 배포용 yml 추가 (#30)

* Feature/add qna/8 (#33)

* Feature : 질문 미리보기 표현 위한 QuestionPreviewDTO 생성 및 구현

Issues : #8

* Feature : 질문에 대한 답변 Entity 생성

Issues : #8

* Feature : 질문 Entity 생성

Issues : #8

* Feature : QuestionController 생성 및 질문 관련 API 추가

Issues : #8

* Feature : AnswerController 생성 및 답변 관련 API 추가

Issues : #8

* Feature : AnswerDTO 생성

Issues : #8

* Feature : QuestionDTO 생성

Issues : #8

* Feature : AnswerRepository 생성

Issues : #8

* Feature : QuestionRepository 생성

Issues : #8

* Feature : QnAService 생성 및 QnA 처리 기능 구현

Issues : #8

* Feature : 회원 기능 구현 + 채팅 기능 구현 + 에러코드 추가 #37 #38 #39 (#43)

* Feature: 채팅방 생성 및 채팅, 채팅 조회 controller 구현

Issues : #37

* Feature : 채팅과 채팅방을 저장하기 위한 entity 생성

Issues : #37

* Feature : 채팅 request와 response 및 room 생성 시 id 반환을 위한 dto 생성

Issues : #37

* Feature: 채팅, 채팅방 레포지토리 생성

Issues : #37

* Feature : 채팅 및 채팅방 생성, 채팅 조회 service 구현

Issues : #37

* Feature : 회원 정보 관련 controller 생성

Issues : #38

* Feature : MemberRepository import 추가

Issues : #38

* Feature : Member에 이름 수정 메소드 추가

Issues : #38

* Feature : MemberDTO와 이름 수정용 DTO 추가

Issues : #38

* Feature : 유저 정보 조회, 정보 수정, 탈퇴 기능 service 구현

Issues : #38

* Feature : SolustionRepository에 fk 를 null로 변경하는 쿼리메소드 추가

Issues : #38

* Feature: 에러코드 추가 및 설정 추가

Issues : #39

* Refactor : 불필요한 log 삭제

Isseus : #39

* Feature : MessageConfig 추가

Issues : #39

* Feature : 에러코드 response 반환을 위한 dto 생성

Issues : #39

* Feature : Exception Advice 추가

Issues : #39

* Feature : Exception class 추가

Issues : #39

* Feature : Exception 적용

Issues : #39

* Refactor: application.yml 수정

* Feature/settings/36 (#44)

* Core : jpa의 ddl-auto 설정을 update로 변경

이를 통해 app 재실행 시에도 이전 데이터를 유지할 수 있습니다.

Issues : #36

* Core : 서비스 정의에 영속 볼륨 매핑을 추가하여, Redis와 MySQL 데이터의 영속성을 보장합니다.

이를 통해 재실행 시에도 이전 데이터를 유지할 수 있습니다.

Issues : #36

* Feature : 회원 조회 수정 탈퇴 수정 #47 (#48)

* Feature : 회원 정보 조회 시 provide와 id도 조회가 가능하도록 파일 변경

Issues : #47

* Feature : 회원 탈퇴 시 설정 추가

 answer 와 question에 member 필드가 null이 될 수 있도록 optional 설정을 추가.

Issues : #47

* Feature : 인스턴스 health 체크용 매핑 생성

배포 시 인스턴스 체크할 때 기본 경로가 돌아가야 하기 때문에 컨트롤러 코드 수정.

Issues : #47

* Feature/qn a member/46 (#50)

* Feature : Exception Class 추가

작성자가 아닌 사람이 질문, 답변을 수정하려고 할 경우 발생시키는 예외 클래스를 추가했습니다.

Issues : #46

* Feature : 에러 코드 추가

작성자가 아닌 사람이 질문, 답변을 수정하려고 할 경우 발생하는 예외의 에러코드와 메세지를 추가했습니다.

Issues : #46

* Feature : Exception Advice 추가

작성자가 아닌 사람이 질문, 답변을 수정하려고 할 경우 발생하는 예외의 Exception Advice를 추가했습니다.

Issues : #46

* Feature : 로그인 체크 기능 추가

질문과 관련해서 로그인 후 사용 가능한 기능 들에 대해 로그인 여부와 작성자인지 여부를 확인하는 기능을 추가했습니다.

Issues : #46

* Feature : 로그인 체크 기능 추가

답변과 관련해서 로그인 후 사용 가능한 기능 들에 대해 로그인 여부와 작성자인지 여부를 확인하는 기능을 추가했습니다.

Issues : #46

* Feature : DB에서 조건에 맞는 미리보기를 가져오기 위한 메서드 추가

조건에 맞는 미리보기를 Page 형태로 가져올 수 있도록 메서드를 추가했습니다.

Issues : #46

* Feature : QnA 처리 중 멤버 확인과, 각종 오류 처리하는 로직 구현

Issues : #46

* Feature : 풀이 가져올 때 로그인 한 사용자 관련 로직 처리

로그인한 사용자가 자신의 풀이만 불러올 수 있도록 수정했습니다.

Issues : #46

* Feature : 질문에 문제 제목도 나타나도록 수정 (#52)

* Feature : 질문 엔티티에 문제 엔티티와의 관계 설정 추가

Question 엔티티 내에서 `problemId` 필드의 타입을 Long에서 Problem 엔티티로 변경하였습니다.
이로 인해 직접적인 ID 참조에서 엔티티 참조로의 관계 정의 방식으로 개선되었습니다.
이를 통해 질문에서 대상 문제의 제목을 바로 확인할 수 있습니다.

Issues : #51

* Feature : ProblemRepository 인터페이스 추가

Issues : #51

* Feature : QuestionDTO에 문제 제목 필드 추가

Issues : #51

* Feature : QuestionPreviewDTO에 문제 제목 필드 추가

Issues : #51

* Feature : QuestionRepositry에서 문제 ID 접근 방식 변경

Issues : #51

* Feature : Question에 문제 제목도 포함하도록 각종 메서드 수정

Issues : #51

* Fix : 배포 후 로그인 시 발생 에러 처리 (#54)

시큐리티 코드 수정
Issues : #53

---------

+++ src/main/java/com/example/challengewithmebe/qna/service/QnAService.java
@@ -1,7 +1,16 @@
 package com.example.challengewithmebe.qna.service;
 
+import com.example.challengewithmebe.global.exception.auth.OwnerOnlyOperationException;
+import com.example.challengewithmebe.global.exception.notExist.NotExistAnswerException;
+import com.example.challengewithmebe.global.exception.notExist.NotExistChallengeException;
+import com.example.challengewithmebe.global.exception.notExist.NotExistMemberException;
+import com.example.challengewithmebe.global.exception.notExist.NotExistQuestionException;
 import com.example.challengewithmebe.member.domain.Member;
+import com.example.challengewithmebe.member.dto.MemberDTO;
 import com.example.challengewithmebe.member.repository.MemberRepository;
+import com.example.challengewithmebe.member.service.MemberService;
+import com.example.challengewithmebe.problem.domain.Problem;
+import com.example.challengewithmebe.problem.repository.ProblemRepository;
 import com.example.challengewithmebe.qna.domain.Answer;
 import com.example.challengewithmebe.qna.domain.Question;
 import com.example.challengewithmebe.qna.dto.AnswerDTO;
@@ -31,10 +40,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class QnAService {
 
+    private final MemberService memberService;
     private final EntityManager entityManager;
     private final MemberRepository memberRepository;
     private final AnswerRepository answerRepository;
     private final QuestionRepository questionRepository;
+    private final ProblemRepository problemRepository;
 
     // 모든 질문 return
     public List<QuestionDTO> allQuestions() {
@@ -43,15 +54,23 @@ public class QnAService {
         return questionDTO;
     }
 
+    // 한 개의 질문과 그에 대한 답변들
+    public QuestionDTO findOneQuestion(Long questionId) {
+        Question q = questionRepository.findById(questionId).orElseThrow(NotExistQuestionException::new);
+        QuestionDTO qDTO = convertToQuestionDTO(q);
+        return qDTO;
+    }
+
     // 특정 문제에 대한 질문 return
     public List<QuestionDTO> questionForOneProblem(Long problemId) {
         List<QuestionDTO> questionDTO =
-                convertToQuestionDTOs(questionRepository.findByProblemIdOrderByModifiedAtDesc(problemId));
+                convertToQuestionDTOs(questionRepository.findByProblemId_IdOrderByModifiedAtDesc(problemId));
             return questionDTO;
     }
 
     // 질문 추가
-    public Long addQuestion(QuestionDTO question) {
+    public Long addQuestion(QuestionDTO question, Long memberId) {
+        question.setMemberId(memberId);
         Question q = questionRepository.save(convertToQuestion(question));
         return q.getId();
     }
@@ -67,60 +86,75 @@ public class QnAService {
     }
 
     // 답변 작성
-    public Long addAnswer(AnswerDTO answerDTO) {
+    public Long addAnswer(AnswerDTO answerDTO, Long memberId) {
+        answerDTO.setMemberId(memberId);
         return answerRepository.save(convertToAnswer(answerDTO)).getId();
     }
 
     // 질문 수정
-    public Long updateQuestion(QuestionDTO questionDTO){
-        Optional<Question> existingQuestionOptional = questionRepository.findById(questionDTO.getId());
-        Question question;
-        if(existingQuestionOptional.isEmpty()){
-            return addQuestion(questionDTO);
+    public Long updateQuestion(QuestionDTO questionDTO, Long memberId){
+        Question q = questionRepository.findById(questionDTO.getId()).orElseThrow(NotExistQuestionException::new);
+        if(q.getMemberId().getId() != memberId){
+            throw new OwnerOnlyOperationException();
         }
-
-        question = existingQuestionOptional.get();
-        question.update(questionDTO);
-        return questionRepository.save(question).getId();
+        q.update(questionDTO);
+        return questionRepository.save(q).getId();
     }
 
     // 답변 수정
-    public Long updateAnswer(AnswerDTO answerDTO){
-        Optional<Answer> existingAnswerOptional = answerRepository.findById(answerDTO.getId());
-        Answer answer;
-        if(existingAnswerOptional.isEmpty()){
-            return addAnswer(answerDTO);
+    public Long updateAnswer(AnswerDTO answerDTO, Long memberId){
+        Answer a = answerRepository.findById(answerDTO.getId()).orElseThrow(NotExistAnswerException::new);
+        if(a.getMemberId().getId() != memberId){
+            throw new OwnerOnlyOperationException();
         }
-
-        answer = existingAnswerOptional.get();
-        answer.update(answerDTO);
-        return answerRepository.save(answer).getId();
+        a.update(answerDTO);
+        return answerRepository.save(a).getId();
     }
 
     // 질문 삭제
-    public boolean deleteOneQuestion(Long questionId) {
+    public boolean deleteOneQuestion(Long questionId,Long memberId) {
+        Question q = questionRepository.findById(questionId).orElseThrow(NotExistQuestionException::new);
+        if(q.getMemberId().getId() != memberId){
+            throw new OwnerOnlyOperationException();
+        }
         questionRepository.deleteById(questionId);
         return true;
     }
 
     // 답변 삭제
-    public boolean deleteOneAnswer(Long answerId) {
+    public boolean deleteOneAnswer(Long answerId, Long memberId) {
+        Answer a = answerRepository.findById(answerId).orElseThrow(NotExistAnswerException::new);
+        if(a.getMemberId().getId() != memberId){
+            throw new OwnerOnlyOperationException();
+        }
         answerRepository.deleteById(answerId);
         return true;
     }
 
     // 모든 질문에 대한 미리보기 return
-    public Page<QuestionPreviewDTO> findAllQuestionPreviews(int page, int size){
+    public Page<QuestionPreviewDTO> findAllQuestionPreviews(int page, int size, String type, Long memberId){
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "modifiedAt"));
-        Page<Question> questions = questionRepository.findAll(pageable);
+        Page<Question> questions;
+        if(type.equals("my")){
+            questions = questionRepository.findByMemberId_Id(memberId, pageable);
+        }else{
+            questions = questionRepository.findAll(pageable);
+        }
+
         Page<QuestionPreviewDTO> previewDTOS = questions.map(this::convertFromQuestionToQuestionPreviewDTO);
         return previewDTOS;
     }
 
     // 특정 문제에 대한 질문 미리보기 return
-    public Page<QuestionPreviewDTO> findQuestionPreviewsForSpecificProblem(Long problemId,int page, int size){
+    public Page<QuestionPreviewDTO> findQuestionPreviewsForSpecificProblem(Long problemId,int page, int size, String type, Long memberId){
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "modifiedAt"));
-        Page<Question> questions = questionRepository.findByProblemId(problemId, pageable);
+        Page<Question> questions;
+        if(type.equals("my")){
+            questions = questionRepository.findByProblemId_IdAndMemberId_Id(problemId, memberId, pageable);
+        }else {
+            questions = questionRepository.findByProblemId_Id(problemId, pageable);
+        }
+
         Page<QuestionPreviewDTO> previewDTOS = questions.map(this::convertFromQuestionToQuestionPreviewDTO);
         return previewDTOS;
     }
@@ -131,26 +165,6 @@ public class QnAService {
             dtos.add(convertToQuestionDTO(q));
         }
         return dtos;
-    }
-
-    // MemberId를 확인하고 null이 아니면 사용자의 정보를, null이면 default를 return
-    private String[] memberCheck(Supplier<Optional<Member>> memberSupplier) {
-        String[] memberInfo = new String[3];
-        Optional<Member> memberOptional = memberSupplier.get(); // Supplier로부터 Optional<Member>를 받아옵니다.
-
-        if (!memberOptional.isPresent()) {
-            // 멤버가 탈퇴한 경우 처리
-            memberInfo[0] = null;
-            memberInfo[1] = "탈퇴한 사용자";
-            memberInfo[2] = "https://default_image.jpg"; // 프로필 이미지 정보가 없으므로 기본 이미지 URL 설정
-        } else {
-            // 멤버가 탈퇴하지 않은 경우
-            Member member = memberOptional.get(); // Optional<Member>에서 Member 인스턴스를 추출
-            memberInfo[0] = String.valueOf(member.getId());
-            memberInfo[1] = member.getName();
-            memberInfo[2] = member.getImgUrl();
-        }
-        return memberInfo;
     }
 
     private QuestionDTO convertToQuestionDTO(Question question) {
@@ -164,9 +178,16 @@ public class QnAService {
                 .map(Long::valueOf)
                 .orElse(null);
 
+        String[] problemInfo = problemCheck(Optional.ofNullable(question.getProblemId()));
+
+        Long problemId = Optional.ofNullable(problemInfo[0])
+                .map(Long::valueOf)
+                .orElse(null);
+
         return QuestionDTO.builder()
                 .id(question.getId())
-                .problemId(question.getProblemId())
+                .problemId(problemId)
+                .problemTitle(problemInfo[1])
                 .title(question.getTitle())
                 .imgUrl(question.getImgUrl())
                 .content(question.getContent())
@@ -179,7 +200,6 @@ public class QnAService {
                 .answers(answerDTOs)
                 .build();
     }
-
 
 
     private AnswerDTO convertToAnswerDTO(Answer answer){
@@ -203,19 +223,13 @@ public class QnAService {
     }
 
     private Question convertToQuestion(QuestionDTO question){
+        Member member = memberRepository.findById(question.getMemberId())
+                .orElseThrow(NotExistMemberException::new);
         Optional<Member> memberOptional = memberRepository.findById(question.getMemberId());
-        if(memberOptional.isEmpty()){
-            return Question.builder()
-                    .problemId(question.getProblemId())
-                    .title(question.getTitle())
-                    .imgUrl(question.getImgUrl())
-                    .content(question.getContent())
-                    .build();
-        }
-
-        Member member = memberOptional.get();
+        Problem problem = problemRepository.findById(question.getProblemId())
+                .orElseThrow(NotExistChallengeException::new);
         return Question.builder()
-                .problemId(question.getProblemId())
+                .problemId(problem)
                 .title(question.getTitle())
                 .imgUrl(question.getImgUrl())
                 .content(question.getContent())
@@ -240,6 +254,7 @@ public class QnAService {
         return QuestionPreviewDTO.builder()
                 .id(questionDTO.getId())
                 .problemId(questionDTO.getProblemId())
+                .problemTitle(questionDTO.getProblemTitle())
                 .title(questionDTO.getTitle())
                 .memberId(questionDTO.getMemberId())
                 .answerCounts(questionDTO.getAnswerCounts())
@@ -253,6 +268,48 @@ public class QnAService {
     private QuestionPreviewDTO convertFromQuestionToQuestionPreviewDTO(Question question){
         QuestionDTO q = convertToQuestionDTO(question);
         return convertToQuestionPreviewDTO(q);
+    }
+
+    private MemberDTO getMemberInfo(Long memberId){
+        return memberService.getInfo(memberId);
+    }
+
+    // MemberId를 확인하고 null이 아니면 사용자의 정보를, null이면 default를 return
+    private String[] memberCheck(Supplier<Optional<Member>> memberSupplier) {
+        String[] memberInfo = new String[3];
+        Optional<Member> memberOptional = memberSupplier.get(); // Supplier로부터 Optional<Member>를 받아옵니다.
+
+        if (!memberOptional.isPresent()) {
+            // 멤버가 탈퇴한 경우 처리
+            memberInfo[0] = null;
+            memberInfo[1] = "탈퇴한 사용자";
+            memberInfo[2] = "https://default_image.jpg"; // 프로필 이미지 정보가 없으므로 기본 이미지 URL 설정
+        } else {
+            // 멤버가 탈퇴하지 않은 경우
+            Member member = memberOptional.get(); // Optional<Member>에서 Member 인스턴스를 추출
+            memberInfo[0] = String.valueOf(member.getId());
+            memberInfo[1] = member.getName();
+            memberInfo[2] = member.getImgUrl();
+        }
+        return memberInfo;
+    }
+
+    private String[] problemCheck(Optional<Problem> problemOptional) {
+        String[] problemInfo = new String[2];
+
+        if(problemOptional.isEmpty()){
+            // 문제가 삭제된 경우
+            problemInfo[0] = null;
+            problemInfo[1] = "존재하지 않는 문제입니다.";
+            return problemInfo;
+        }
+
+        Problem p = problemOptional.get();
+        problemInfo[0] = String.valueOf(p.getId());
+        problemInfo[1] = p.getTitle();
+        return problemInfo;
+
+
     }
 
 }
